# niakofa-repo/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Niakofa - Map-First Community Help Platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: User not found
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a user account (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Account deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  message:
                    type: string
        "400":
          description: Invalid id
        "500":
          description: Failed to delete account
...
components:
  schemas:
...
    User:
      type: object
      required: [id, name, email, is_helper, helper_mode_active, help_count, benevolence_wallet, goodwill_score]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        avatar_url:
          type: ["string", "null"]
        is_helper:
          type: boolean
        helper_mode_active:
          type: boolean
        lat:
          type: ["number", "null"]
        lng:
          type: ["number", "null"]
        trust_score:
          type: ["number", "null"]
        help_count:
          type: integer
        neighborhood:
          type: ["string", "null"]
        benevolence_wallet:
          type: number
          description: Accumulated helper earnings and contributions
        goodwill_score:
          type: integer
          description: Community reputation score
        no_show_count:
          type: integer
          description: Number of times helper released a claimed request (no-show tracking)
          default: 0
        is_admin:
          type: boolean
          description: RBAC admin flag — see requireAdmin() middleware
        created_at:
          type: string
        account_type:
          type: ["string", "null"]
          enum: [individual, organization, business, sponsor, null]
        city:
          type: ["string", "null"]
        specialties:
          type: ["array", "null"]
          items:
            type: string
        community_id:
          type: ["integer", "null"]
          description: The county/community pool this user is scoped to — resolves the livable-wage floor and pool-health multiplier for this user. Null = legacy global/unassigned bucket.
        helper_status:
          type: ["string", "null"]
          enum: [offline, pending, approved, denied, null]
          description: Helper application/approval state. "offline" is the default for accounts that have never applied. Only "approved" accounts can set helper_mode_active to true (PATCH /users/{id}/helper-mode enforces this server-side).
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        avatar_url:
          type: string
        neighborhood:
          type: string
        is_helper:
          type: boolean
        city:
          type: string
        specialties:
          type: array
          items:
            type: string
        phone_masked:
          type: string
        quick_replies:
          type: array
          items:
            type: string
...